internal: Split up the api.go file
This file was a remnant of the old internal implementation for the API
access. Only two unrelated functionalities were left and so they have been moved
into their own files.
